refactor: move FareCalculator to application
