+++ src/main/java/nextstep/subway/application/dto/PathResponse.java
@@ -5,8 +5,8 @@ import java.util.stream.Collectors;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 
+import nextstep.subway.application.FareCalculator;
 import nextstep.subway.domain.model.Path;
-import nextstep.subway.domain.model.PathType;
 import nextstep.subway.domain.model.Station;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
